+++ src/form-elements/FormElementABN.tsx
@@ -248,6 +248,7 @@ function FormElementABN({
             value={value}
             validationMessage={validationMessage}
             lookupButtonConfig={element.lookupButton}
+            allowLookupOnEmptyValue={element.runLookupOnClear}
           />
         </div>
         {(((isDirty || displayValidationMessage) &&