+++ src/form-elements/FormElementEmail.tsx
@@ -65,6 +65,7 @@ function FormElementEmail({
             value={value}
             validationMessage={validationMessage}
             lookupButtonConfig={element.lookupButton}
+            allowLookupOnEmptyValue={element.runLookupOnClear}
           />
         </div>