+++ src/form-elements/FormElementTelephone.tsx
@@ -65,6 +65,7 @@ function FormElementTelephone({
             value={value}
             validationMessage={validationMessage}
             lookupButtonConfig={element.lookupButton}
+            allowLookupOnEmptyValue={element.runLookupOnClear}
           />
         </div>