+++ CHANGELOG.md
@@ -11,9 +11,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Autocomplete element not clearing on lookup returning undefined on value
 
+### Added
+
+- support for `runLookupOnClear` to allow running lookups on empty values
+
 ### Changed
 
 - blur autocomplete when an option is clicked
+- form element to display options as soon as possible instead of wait for all dynamic options to be loaded
 
 ## [4.2.0] - 2023-07-12